+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## unreleased
 
+- feat: Bring back SentryOptions.enabled #818
 - fix: Add eventId to user feedback envelope header #809
 - feat: Manually capturing User Feedback #804